Pin CI tool versions and environments, and make the silent gates actually verify (#1)
CI had been red since 2026-07-20, and not because of our code. Fixing
the config exposed a chain of defects that had each been hidden by the
one in front of it.

- ruff was unpinned in CI (`uvx ruff` = always latest) while local was
  0.15.22; ruff 0.16.0 widened its default rule set from 118 rules to
  413, so the repo could go red with no commit of ours. Pinned, plus a
  root ruff.toml unifying target-version and stopping config discovery
  inside the repo. All 42 findings fixed.
- The secret scan died on checkout's shallow clone and reported
  `scanned ~0 bytes` — while concluding green. Fixed with fetch-depth: 0.
- data-contracts and schema-validate had been skipped for 8 days behind
  `needs: python`. Both failed once revived, both for real reasons: the
  entity schema had fallen behind entities.json (debut/agency/wd_id),
  and the snapshot gate was validating signal-series documents against
  the snapshot schema.
- pyright had never run at all — ruff died ahead of it. Pinning the
  version was not enough; the environment matters too, so the step now
  installs the deps that are actually required at module scope.

Both gates now fail when they match zero files. A green gate that
verified nothing is worse than a red one.

Recorded as D-030.

+++ modules/sonic-profile/src/sonic_profile/features.py
@@ -39,8 +39,8 @@ def _pulse_clarity(onset_env: np.ndarray, sr: int, hop: int) -> float:
         return 0.0
     ac = ac / ac[0]
     # 30~300 BPM에 해당하는 lag만 본다(그 밖의 주기는 박이 아님)
-    lo = max(1, int(round(60.0 / 300.0 * sr / hop)))
-    hi = min(ac.size - 1, int(round(60.0 / 30.0 * sr / hop)))
+    lo = max(1, round(60.0 / 300.0 * sr / hop))
+    hi = min(ac.size - 1, round(60.0 / 30.0 * sr / hop))
     if hi <= lo:
         return 0.0
     return float(np.clip(np.max(ac[lo : hi + 1]), 0.0, 1.0))
@@ -48,7 +48,7 @@ def _pulse_clarity(onset_env: np.ndarray, sr: int, hop: int) -> float:
 
 def _key_mode(y: np.ndarray, sr: int) -> tuple[str, str, float]:
     """chroma → (key, mode, 상관계수). 정확도가 낮아 집계 전용(RULES §3)."""
-    import librosa
+    import librosa  # type: ignore  # 선택적 중량 의존성 (CI 타입체크 환경에 없음)
 
     chroma = librosa.feature.chroma_cqt(y=y, sr=sr)
     profile = np.mean(chroma, axis=1)
@@ -66,7 +66,7 @@ def _key_mode(y: np.ndarray, sr: int) -> tuple[str, str, float]:
 
 def extract(y: np.ndarray, sr: int = SR, *, low_hz: float = LOW_HZ_DEFAULT) -> dict[str, Any]:
     """오디오 배열 → RULES §3 지표 dict. 미해석 입력은 Unresolved를 던진다."""
-    import librosa
+    import librosa  # type: ignore  # 선택적 중량 의존성 (CI 타입체크 환경에 없음)
 
     if y.ndim > 1:
         y = librosa.to_mono(y)
@@ -123,7 +123,7 @@ def extract(y: np.ndarray, sr: int = SR, *, low_hz: float = LOW_HZ_DEFAULT) -> d
 
 def engine_provenance(low_hz: float = LOW_HZ_DEFAULT) -> dict[str, Any]:
     """값의 일부인 엔진 설정 — 바뀌면 시리즈를 버전 분리해야 한다(RULES §2)."""
-    import librosa
+    import librosa  # type: ignore  # 선택적 중량 의존성 (CI 타입체크 환경에 없음)
 
     return {
         "engine": "librosa",
